Make /device/ip_address required
The request builder constructor now has a required param of Device
and the Device builder constructor requires the IP address. If we
drop this requirement in the future, we can always add aditional
constructors. This ensures that the problem is caught as compile
time.

Also reformatted tests as they were a bit messy.