Stop using Rewrite and update Checkstyle
I found Rewrite doesn't actually do that great of job at matching the
Checkstyle config and just leads to more confusion.

This also updates Checkstyle, makes it run during the test phase,
increases the severity of warnings so that they fail, and inlines the
version numbbers (to make things simpler for tooling).

It also renames the suppression file to make it clearer what it is for.